add namespaces data
